requirements.txt
keboola.component==1.4.4
unidecode
pyzoom==1.0.5
mock==5.0.2
freezegun==1.2.2
keboola.csvwriter == 1.0.1
